add a multi-purpose file download utility function and tests, remove individual implementations
